+++ CHANGELOG.md
@@ -9,8 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add event TUI editor, `event new` and `event edit` now work without arguments to open the TUI.
+- Add event TUI editor, `event new` and `event edit` now work without arguments to open the TUI
 - `Percent complete` now accepts only numeric input in todo editor
+
+### Changed
+
+- Status is now required for drafts since it has a default value
 
 ### Fixed